fixed problems in Linux side
