+++ src/nupic/algorithms/Anomaly.cpp
@@ -28,7 +28,6 @@
 
 #include "nupic/algorithms/Anomaly.hpp"
 #include "nupic/utils/Log.hpp"
-#include "nupic/utils/MovingAverage.hpp"
 #include "nupic/types/Sdr.hpp"
 
 using namespace std;
@@ -121,26 +120,6 @@ Real Anomaly::compute(vector<UInt>& active, vector<UInt>& predicted, int timesta
 
   return score;
 }
-
-template<class Archive>
-void Anomaly::save_ar(Archive & ar) const {
-  std::string name("Anomaly");
-  ar(CEREAL_NVP(name), 
-      CEREAL_NVP(mode_), 
-      CEREAL_NVP(binaryThreshold_), 
-      CEREAL_NVP(likelihood_),
-      CEREAL_NVP(movingAverage_));
-}
-template<class Archive>
-void Anomaly::load_ar(Archive & ar) {
-  std::string name;
-  ar(CEREAL_NVP(name),
-      CEREAL_NVP(mode_),
-      CEREAL_NVP(binaryThreshold_));
-  ar(CEREAL_NVP(likelihood_));
-  ar(CEREAL_NVP(movingAverage_));
-}
-
 
 bool Anomaly::operator==(const Anomaly &a) const {
   if (mode_ != a.mode_) return false;
